requirements.txt
# AI Tutor Backend - PDF Storage & Vector Database Requirements

# Core Framework
fastapi>=0.100.0
uvicorn>=0.20.0

# Database & Vector Storage  
supabase>=2.0.0
# psycopg2-binary>=2.9.5  # Skip for now, will use supabase client only
# pgvector>=0.2.0  # Skip for now, pgvector is enabled in Supabase directly

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0
python-multipart>=0.0.5

# AI & Embeddings
openai>=1.0.0
tiktoken>=0.5.0

# Async Support
aiofiles>=0.20.0
httpx>=0.25.0

# Data Processing
numpy>=1.20.0
pandas>=1.5.0

# Utilities
python-slugify>=8.0.0
pillow>=10.0.0

# Development & Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0